==> notebooks_cleaning/1_rename_columns.ipynb <==
import sys
sys.path.insert(0, '..')

import pandas as pd
from helpers import *

df_raw = pd.read_excel(path_data + "IOW - _Cleaned and anonymised.xlsx", skiprows=1)

rename = read_json(path_json + "rename_columns.json")

def generate_output():

    df_rename = df_raw.rename(columns=rename)
    df_final = df_rename.iloc[1:] # First row contains column information that is not needed

    df_final.to_excel("../data/IOW - 1 - Renamed Columns.xlsx", index=False)

    return df_final

generate_output().head(2)

# def generate_question_index():

#     df_question = pd.concat([pd.DataFrame([rename]).T, df_raw.head(4).T], axis=1).reset_index()
#     df_question.columns = ["fulltext", "renamed", "question_type", "example_1", "example_2", "example_3"]

#     df_question.to_excel("../data/Question_Index.xlsx", index=False)

#     return df_question

# generate_question_index().head(2)
