==> notebooks_cleaning/2_rename_other.ipynb <==
import sys
sys.path.insert(0, '..')

import pandas as pd
from helpers import *

df = pd.read_excel("../data/IOW - 1 - Renamed Columns.xlsx")

other_cols = ["WORD_person", "PERSON_relationship"]

# Read the json files in the json folder as dictionaries

rename = {
    c:invert_one_to_many_dict( read_json(path_json + f"rename_{c}_other.json") ) for c in other_cols
}

## 1 Apply the hand-coded "Other" column classification to the main column

def rename_other(df, column_name):
    
    other_column_name = f"{column_name}_other"

    df.loc[~df[other_column_name].isna(), column_name] = df[other_column_name].dropna().str.lower().str.strip().replace(rename[column_name])
    return df

result = rename_other(df,"WORD_person")

print("Distribution of Categories after grouping 'Others'")
result["WORD_person"].value_counts()

result = rename_other(df, "PERSON_relationship")

print("Distribution of Categories after grouping 'Others'")
result["PERSON_relationship"].value_counts()

## 2 Drop "Other" columns

result = result.drop(columns=["WORD_person_other", "PERSON_relationship_other", "PERSON_gender_other"])

## Export

result.to_excel(path_data + "IOW - 2 - Renamed Others.xlsx", index=False)
