==> notebooks_cleaning/3_clean_text_columns.ipynb <==
import sys
sys.path.insert(0, '..')

import pandas as pd
import numpy as np

import pandas as pd
import nltk
import re

from helpers import *

df = pd.read_excel(path_data + "IOW - 2 - Renamed Others.xlsx")

text_columns = [
    "WORD_association",
    "PERSON_association",
    "BRAND_association",   
    "SELF_association"
]

## 1 Goal: Remove Sentences (Invalid Responses)

### 1.0 Lower and Strip

# lower and strip 

df[text_columns] = df[text_columns].apply(lambda x: x.str.lower().str.strip())

### 1.1 Remove: Entries with Contractions

These are likely to be sentences, hence not of interest to us

def remove_contractions(series):
    
    expr = "'t |'m |'re |'ve "

    invalid = series.str.contains(expr).fillna(False)

    print(series.name)
    print(series.loc[invalid].value_counts())
    print(" ")

    series.loc[invalid] = np.nan

    return series

df[text_columns] = df[text_columns].apply(remove_contractions)

### 1.2 Remove: Entries with first person singular

Also typical in sentences

def remove_first_person(series):

    invalid = (" " + series).str.contains(" i ").fillna(False)

    print(series.name)
    print(series[invalid].value_counts())
    print(" ")

    series.loc[invalid] = np.nan

    return series

df[text_columns] = df[text_columns].apply(remove_first_person)

## 2 Goal: Clean up special characters

### 2.1 Use unified seperator (,)

def replace_alternate_with_comma(series):

    no_comma = ~series.str.contains(",").fillna(False)
    series.loc[no_comma] = series.loc[no_comma].str.replace(" ", ",").str.replace(";", ",")

    return series

df[text_columns] = df[text_columns].apply(replace_alternate_with_comma)

### 2.2 Replace special characters, remove unnecessary white spaces and unify ngram format

def clean_text(text):
    text = re.sub(r'\s+', ' ', text) # Replace multiple spaces with one
    text = re.sub(r',+', ',', text) # replace multiple commas with one
    text = re.sub(r'[^a-zA-Z0-9&,:\-]', '', text) # remove special characters, allowing only & and , and : and -

    text = text.replace(",", " ").strip().replace(" ", ",") # Remove commas at the start and end of text

    return text

def clean_word(word):
    word = word.strip() # remove spaces at the start and end of word

    word = " ".join([part.strip() for part in word.split("-")]) # remove spaces between hyphenated words
    word = " ".join([part.strip() for part in word.split(":")]) # remove spaces between colonated words

    word = word.replace(" ", "-") # replace spaces with hyphens (for ngrams)

    return word

def clean(series):
    series = series.apply(clean_text)
    series = series.str.split(",").apply(lambda array: ",".join([clean_word(x) for x in array]))
    
    return series

df[text_columns] = df[text_columns].fillna("").apply(clean)

## 3 Goal: Standardize words

### 3.1 Correct typos

#pip install pyspellchecker

from spellchecker import SpellChecker

df_allwords = df[text_columns].fillna("")\
                .apply(lambda x: ",".join(x), axis=1)\
                    .str.split(",").explode()

spell = SpellChecker()

unknown_words = spell.unknown(df_allwords.value_counts().index)

# ngrams are marked as unknown, however the majority of errors are within unigrams, hence we will focus only on these for simplicity

unknown_unigrams = [word for word in unknown_words if len(word.split("-")) == 1] 

# Are there any preferred ones to fix?

df_allwords.value_counts()[unknown_unigrams]

# Note:
# 
# df_allwords.unique() list was fed to gpt-4 to get the correct spelling and copied into typos.json

typo_dict = read_json(path_json + "typos.json")

def correct_typo(series):

    def correct_typo_word(word):
        if word in typo_dict:
            return typo_dict[word]
        else:
            return word

    series = series.apply(lambda x: apply_wordlist(x, correct_typo_word))
    
    return series

df[text_columns] = df[text_columns].apply(correct_typo)

### 3.2 Convert some nouns to adjectives

noun_to_adj_dict = read_json(path_json + "noun_to_adj.json")

def noun_to_adj(series):

    return series.apply(lambda x: apply_wordlist(x, lambda y: noun_to_adj_dict[y] if y in noun_to_adj_dict else y))

df[text_columns] = df[text_columns].apply(noun_to_adj)

# for i in range(0, len(vc)):
#     word = vc.index[i]
#     count = vc[i]
    
#     if count > 1:
#         print(word)

### 3.3 Plurals to Singular

import inflect

p = inflect.engine()

def plural_to_singular(series):

        def plural_to_singular_word(word):
            if len(word) > 0:
                if p.singular_noun(word):
                        return p.singular_noun(word)
                else:
                        return word
            else: 
                return word

        return series.apply(apply_wordlist, args=(plural_to_singular_word,))


df[text_columns] = df[text_columns].apply(plural_to_singular)

## Final: Export

df.to_excel(path_data + "IOW - 3 - Cleaned Text.xlsx", index=False)
