==> notebooks_cleaning/4_clean_remaining_columns.ipynb <==
import sys
sys.path.insert(0, '..')

import numpy as np
import pandas as pd

from helpers import *

df = pd.read_excel(path_data + "IOW - 3 - Cleaned Text.xlsx")
q = pd.read_excel(path_data + "Question_Index.xlsx")

## 1 SERVICE Replace misssing data

(df[q[q.source == "service"].renamed.tolist()] == "DATA_EXPIRED").mean().apply("{:.2%}".format)

df.loc[:, q[q.source == "service"].renamed.tolist()] = df.loc[:, q[q.source == "service"].renamed.tolist()].replace("DATA_EXPIRED", np.nan)

## 2 SPECIFIC Enrich PERSON_gender 

def transfer_gender(row):

    if row.PERSON_relationship == "Wife or girlfriend":
        return "Female"

    elif row.PERSON_relationship == "Sister":
        return "Female"

    elif row.PERSON_relationship == "Husband or boyfriend":
        return "Male"

    elif row.PERSON_relationship == "Brother":
        return "Male"

    else:
        return row.PERSON_gender

df.PERSON_gender = df[["PERSON_relationship", "PERSON_gender"]].apply(transfer_gender, axis=1)

df.PERSON_gender.isna().sum()

## 3 Nominal Columns

nominal_columns = q[q.question_type == "nominal"].renamed.tolist()

pd.concat([
    df[column].value_counts().reset_index() for column in nominal_columns
],axis=1).fillna("-").to_excel(path_data + "ColumnInfo - Nominal.xlsx")

"BRAND_name", 
                

"PERSON_relationship", 
                

"PERSON_gender", 
                

"QUESTION_trust", 
                

"languages", 
                

"age", 
                

"sex", 

"ethnicity", 
                

"country_of_birth", 
                

"country_of_residence", 
                

"nationality",
                

"first_language",
                

"is_student",
                

"employment"

qs = pd.read_excel(path_data + "Question_Index.xlsx")

qs[["renamed", "question_type"]]
